Final Copy
Final copy:
Loop wasn't working as intended in area_perimeter() and shape_type(). If an invalid input was entered into the questions, the code would recognise that the user input was invalid but would still save the first input to use for further lines of the code
The same happened with shape_type(), if an invalid input was enter into the question, the code would recognise that the user input was invalid but it wouldn't go onto the next step of calculating.